% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare.R
\name{delta}
\alias{delta}
\title{Create a Delta instance representing a time delta.}
\usage{
delta(seconds = 0, minutes = 0, hours = 0, days = 0, weeks = 0,
  months = 0, years = 0)
}
\value{
An instance of Delta.
}
\description{
Create a Delta instance representing a time delta.
}
