% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare.R
\name{select}
\alias{select}
\title{Initiate a select statement}
\usage{
select(start = NULL, end = NULL)
}
\arguments{
\item{start}{A POSIXlt date.}

\item{end}{A POSIXlt date.}
}
\value{
An instance of Select bound to start/end.
}
\description{
Initiate a select statement
}
