% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{class}
\name{Sentenai-class}
\alias{Sentenai-class}
\alias{Sentenai}
\title{Sentenai API client.}
\description{
Sentenai API client.
}
\section{Fields}{

\describe{
\item{\code{auth_key}}{Valid API authentication key.}

\item{\code{host}}{(optional) Defaults to https://api.sentenai.com}
}}


