% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sentenaiWithKey}
\alias{sentenaiWithKey}
\title{Create Sentenai API client.}
\usage{
sentenaiWithKey(auth_key, host = "https://api.sentenai.com")
}
\arguments{
\item{auth_key}{Valid API authentication key.}

\item{host}{(optional) Defaults to https://api.sentenai.com}
}
\description{
Create Sentenai API client.
}
