% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare.R
\name{stream}
\alias{stream}
\title{Create a Stream instance}
\usage{
stream(name, filter = NULL)
}
\arguments{
\item{name}{Name of the Sentenai stream.}

\item{filter}{Conditions applied to the raw stream.}
}
\value{
An instance of Stream.
}
\description{
Create a Stream instance
}
